Default feedbackStategy to Always and serverErrors to null
As agreed in #21, `feedbackStrategy` needs a default value and `Always`
is a very understandable default. So this PR obsoletes #21 which has
other unwanted changes too.

Also, `serverErrors` is very frequently passed as `null` so I made that
a default value too.

I removed explicit values for these props in tests that seemed unrelated
to the props in question.

Test plan: `yarn test && yarn flow`